Add query param validation to remaining new-style services [azuredevops appveyor npm]
Remove now-obsolete code.

Close #2675
