+++ core/base-service/route.js
@@ -18,10 +18,8 @@ const isValidRoute = Joi.object({
     then: Joi.array().items(Joi.string().required()),
   }),
   queryParamSchema: Joi.object().schema(),
-  queryParams: Joi.array().items(Joi.string().required()),
 })
   .xor('pattern', 'format')
-  .oxor('queryParamSchema', 'queryParams')
   .required()
 
 function assertValidRoute(route, message = undefined) {
@@ -71,12 +69,12 @@ function namedParamsForMatch(captureNames = [], match, ServiceClass) {
   return result
 }
 
-function getQueryParamNames({ queryParams = [], queryParamSchema }) {
+function getQueryParamNames({ queryParamSchema }) {
   if (queryParamSchema) {
     const { children, renames = [] } = Joi.describe(queryParamSchema)
     return Object.keys(children).concat(renames.map(({ from }) => from))
   } else {
-    return queryParams
+    return []
   }
 }
